requirements.txt
streamlit>=1.18.0
Pillow==9.1.0
pytesseract==0.3.9
opencv-python-headless==4.5.5.64
numpy>=1.23.5,<2.0.0
